requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
supabase==2.4.1
python-dotenv==1.0.1